pyproject.toml
[project]
name = "freqtrade-strategies"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Mao Bui", email = "maonguyen199873@gmail.com"},
]
dependencies = [
    "ruff>=0.6.4",
    "tqdm>=4.66.5",
    "freqtrade>=2024.8",
    "langchain>=0.2.16",
    "langchain-openai>=0.1.23",
    "langchain-community>=0.2.16",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
